Move app left nav out to its own component
